yarn test, prettier:format

+++ src/app.ts
@@ -1,10 +1,14 @@
-import express, { type Express } from 'express';
-import cors from 'cors';
-import * as dotenv from 'dotenv';
-import db from './models/index';
-import authRouter from './routes/auth.routes';
+import express, { type Express } from "express";
+import cors from "cors";
+import * as dotenv from "dotenv";
+import db from "./models/index";
+import authRouter from "./routes/auth.routes";
 import userRouter from "./routes/user.routes";
-import session from 'express-session';
+import session from "express-session";
+import * as process from "process";
+
+const initIndex = process.argv.indexOf("--init=true");
+const init = initIndex !== -1;
 
 const app: Express = express();
 
@@ -13,27 +17,29 @@ dotenv.config();
 app.use(cors());
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
-app.use(session({
-    secret: 'your-secret-key-here',
+app.use(
+  session({
+    secret: "your-secret-key-here",
     resave: false,
     saveUninitialized: true,
-}));
+  })
+);
 
-app.get('/', (_req, res) => {
-    res.send('Welcome to the VirtuaYou UserAuth API.')
-})
+app.get("/", (_req, res) => {
+  res.send("Welcome to the VirtuaYou UserAuth API.");
+});
 
 // database
 const Role = db.role;
 
-//if (init) {
-    db.sequelize.sync({force: true}).then(() => {
-        console.log('Drop and Resync Db');
-        initial();
-    });
-//} else {
-//    db.sequelize.sync();
-//}
+if (init) {
+  db.sequelize.sync({ force: true }).then(() => {
+    console.log("Drop and Resync Db");
+    initial();
+  });
+} else {
+  db.sequelize.sync();
+}
 
 // routes
 app.use(authRouter);
@@ -41,25 +47,25 @@ app.use(userRouter);
 
 // create reference role objects
 function initial() {
-    Role.create({
-        id: 1,
-        name: "owner",
-    });
+  Role.create({
+    id: 1,
+    name: "owner",
+  });
 
-    Role.create({
-        id: 2,
-        name: "agent",
-    });
+  Role.create({
+    id: 2,
+    name: "agent",
+  });
 
-    Role.create({
-        id: 3,
-        name: "monitor",
-    });
+  Role.create({
+    id: 3,
+    name: "monitor",
+  });
 
-    Role.create({
-        id: 4,
-        name: "admin",
-    });
+  Role.create({
+    id: 4,
+    name: "admin",
+  });
 }
 
 export default app;
